Move all BaseHook usages to version_compat in Amazon
