+++ providers/amazon/tests/unit/amazon/aws/operators/test_base_aws.py
@@ -22,11 +22,7 @@ import pytest
 
 from airflow.providers.amazon.aws.hooks.base_aws import AwsBaseHook
 from airflow.providers.amazon.aws.operators.base_aws import AwsBaseOperator
-
-try:
-    from airflow.sdk import BaseHook
-except ImportError:
-    from airflow.hooks.base import BaseHook  # type: ignore[attr-defined,no-redef]
+from airflow.providers.amazon.version_compat import BaseHook
 from airflow.utils import timezone
 
 TEST_CONN = "aws_test_conn"
